requirements.txt
pyyaml
sklearn
boto3
skater
pandas
graphviz
pydot 
matplotlib
plotly